Zecdata_project/requirements.txt
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
countryinfo==0.1.2
cx-Oracle==8.3.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.0.6
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==0.8.3
face-recognition-models==0.3.0
Faker==15.3.3
fastjsonschema==2.15.3
filelock==3.8.0
Flask==2.2.2
graphene==3.2.1
graphene-django==3.0.0b7
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httpie==3.2.1
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
monk==0.13.2
multidict==6.0.4
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
panda==0.3.1
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
requests-toolbelt==0.10.1
rich==13.0.0
scipy==1.9.0
selenium==4.6.1
Send2Trash==1.8.0
six==1.16.0
sketchpy==0.0.23
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
svg.path==6.2
svgpathtools==1.5.1
svgwrite==1.4.3
terminado==0.15.0
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
trio==0.22.0
trio-websocket==0.9.2
turtle==0.0.1
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.16.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wsproto==1.2.0
